move scripts back into script dir and format with ruff
