+++ scripts/compute_products.py
