+++ scripts/compute_scores.py
